Fix deduplication for 3+ duplicates
We now suffix an extra "D" for every subsequent duplicate.

Doing things this way has the nice property of transactions
"mapping to themselves". Parsing the same day of transactions twice
will result in the same set of duplicate transactions. This makes
parsing safe when parsing different statement that have an overlap
in the data they contain (e.g. 3 month and 6 month statement will
both contain a date with the same set of transactions for that date).