Trigger CLI from a binary
Still hitting issues installing the CLI with the entrypiont being a
typescript file - the artifact doesn't exist until after postinstall so the
install step that sets up permissions fails when trying to make the script
executable.

This change attempts to work around this by having the script entrypoint be a binary which can be copied over before the chmod step runs.