Bundle output using webpack
The output of the build is now a single js file for the library.

This was done to make consuming the library easier (no need to worry about
multiple files). Webpack can also be used in future to optimize the build.

The tradeoff with this is that we'll need to pull the CLI into its own
repo.